+++ extensions/memory-core/src/memory/manager-search-orchestration.test.ts
@@ -2,8 +2,8 @@
 import fs from "node:fs/promises";
 import path from "node:path";
 import type { DatabaseSync } from "node:sqlite";
+import { createDeferred } from "openclaw/plugin-sdk/extension-shared";
 import { describe, expect, it, vi } from "vitest";
-import { createDeferred } from "../../../../test/helpers/promise.js";
 import type { EmbeddingProvider } from "./embeddings.js";
 import {
   createManagerIndexFixture,
